% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package_data.R
\name{create_geog_level_lookup}
\alias{create_geog_level_lookup}
\title{Create geographic_level API shorthand to natural language look-up}
\usage{
create_geog_level_lookup()
}
\description{
Script for updating the package data file geog_level_lookup.rda with the look-up between API
shorthands and natural language versions of geographic levels.
}
\examples{
\dontrun{
create_geog_level_lookup()
}
}
\keyword{internal}
