% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ees_meta.R
\name{generate_ees_meta}
\alias{generate_ees_meta}
\title{Generate EES meta}
\usage{
generate_ees_meta(api_meta)
}
\arguments{
\item{api_meta}{}
}
\value{
data frame
}
\description{
This function generates a basic meta data frame approximating what analysts upload alongside
data files to the platform
}
\examples{
get_meta(example_id()) |> generate_ees_meta()
}
