% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{geog_level_lookup}
\alias{geog_level_lookup}
\title{Look-up for API geographic_level shorthands}
\format{
\subsection{\code{geog_level_lookup}}{

A data frame with 18 rows and 2 columns:
\describe{
\item{api_friendly}{API geographic level shorthands}
\item{human_friendly}{Natural language geographic levels}
}
}
}
\source{
curated by explore.statistics@education.gov.uk
}
\usage{
geog_level_lookup
}
\description{
A look-up of API geographic_level shorthands and their respective natural forms.
}
\keyword{datasets}
