% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_versions.R
\name{get_data_versions}
\alias{get_data_versions}
\title{Get publications}
\usage{
get_data_versions(
  dataset_id,
  ees_environment = NULL,
  api_version = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-data-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"}

\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
Data frame listing all available versions of the dataset
}
\description{
Get publications
}
\examples{
get_publications()
}
