% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_versions.R
\name{get_dataset_versions}
\alias{get_dataset_versions}
\title{Get data set versions}
\usage{
get_dataset_versions(
  dataset_id,
  detail = "light",
  ees_environment = NULL,
  api_version = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"}

\item{detail}{Level of detail to return. Given as a character string, it should be one of:
"light" (default) or "full".
Using "light" gives the following:
\itemize{
\item version - version number
\item type - version type (major or minor)
\item total_rows - total number of rows in the listed version
\item release_date - release date of the data set version
\item release_name - name of the release in which the data set version was published
\item time_period_start - first available time period in the data
\item time_period_start - latest available time period in the data
}

Using "full" adds the following in addition to the above:
\itemize{
\item version_id - unique ID code for the data set version
\item status - publication status of the data set version
\item geographic_levels - geographic levels provided
\item filters - names of filters in the data set version
\item indicators - names of indicators in the data set version
\item notes - any additional information
}}

\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
A data frame with X columns:

Data frame listing all available versions of the given data set
}
\description{
Get data set versions
}
\examples{
get_dataset_versions(dataset_id = example_id(group = "attendance"))
}
