% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_geographic_level_codes}
\alias{parse_geographic_level_codes}
\title{Parse API geographic_levels}
\usage{
parse_geographic_level_codes(geographic_levels, verbose = FALSE)
}
\arguments{
\item{geographic_levels}{Vector of API returned geographic_levels}

\item{verbose}{Run in verbose mode with debugging messages}
}
\value{
Data frame of expanded geographic levels
}
\description{
The API generally returns abbreviated and acronym versions of geographic_levels in it's base
output. This function converts those to more hum-readable versions.
}
\examples{
c("NAT", "NAT", "REG", "LA", "LAD") |>
  eesyapi:::parse_geographic_level_codes()
}
\keyword{internal}
