% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_time_codes}
\alias{parse_time_codes}
\title{Parse API time codes}
\usage{
parse_time_codes(time_periods, verbose = FALSE)
}
\arguments{
\item{time_periods}{data frame of API returned time periods and indicators}

\item{verbose}{Run in verbose mode with debugging messages}
}
\value{
Data frame of expanded time codes
}
\description{
The API returns some abbreviated versions of time periods in it's base
output. This function converts those to more human-readable versions.
}
\examples{
data.frame(
  code = c("W1", "W12", "Academic year"),
  period = c("2024", "2025", "202425")
) |>
  eesyapi:::parse_time_codes()
}
\keyword{internal}
