% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset_utils.R
\name{parse_todf_geographies}
\alias{parse_todf_geographies}
\title{Parse geographies for data set query}
\usage{
parse_todf_geographies(geographies)
}
\arguments{
\item{geographies}{String, vector or data frame containing the geographic levels and
locations to be queried.}
}
\value{
String containing json form query for geographies
}
\description{
Create a json query sub-string based on location constraints
}
\examples{
parse_todf_geographies(c("NAT", "REG")) |>
  cat()
parse_todf_geographies(c("NAT|id|dP0Zw", "REG|id|rg3Nj")) |>
  cat()
parse_todf_geographies(c("NAT|id|dP0Zw", "REG")) |>
  cat()
parse_todf_geographies(c("NAT|id|dP0Zw", "REG")) |>
  cat()
}
