% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_dataset_version}
\alias{validate_dataset_version}
\title{Validate data set version}
\usage{
validate_dataset_version(dataset_version)
}
\arguments{
\item{dataset_version}{Chosen data set version}
}
\value{
Logical
}
\description{
Ensures that the data set version is in the correct format, either as a
string following the 'major.minor.patch' format (with optional '*'
wildcards), or as a numeric value.
}
\details{
Numeric values are kept for backwards compatibility, but expect strings will
be more commonly used.
}
\examples{
eesyapi:::validate_dataset_version("2.0.0")
}
\keyword{internal}
