% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_no_rows.R
\name{warning_no_rows}
\alias{warning_no_rows}
\title{Warn on zero rows returned}
\usage{
warning_no_rows(api_result)
}
\arguments{
\item{api_result}{Output from an API get query}
}
\value{
Original input (api_result) unchanged
}
\description{
Warn on zero rows returned
}
\examples{
response_page <- httr::GET(api_url("get-publications", search = "bob")) |>
  httr::content("text") |>
  jsonlite::fromJSON() |>
  eesyapi:::warning_no_rows()
}
\keyword{internal}
