% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset_utils.R
\name{todf_geographies}
\alias{todf_geographies}
\title{Standardise geography inputs}
\usage{
todf_geographies(geographies)
}
\arguments{
\item{geographies}{String, vector, list or data frame containing the geographic levels and
locations to be queried.}
}
\value{
data.frame containing standardised geography specification
}
\description{
Create a standard data frame containing geographic_level and location for use by
query_dataset().
}
\examples{
eesyapi:::todf_geographies(c("NAT", "REG"))
eesyapi:::todf_geographies(c("NAT|code|E92000001", "REG|code|E12000001"))
eesyapi:::todf_geographies(c("NAT|code|E92000001", "REG"))
eesyapi:::todf_geographies(
  list(
    geographic_level = c("REG", "LA"),
    location = c("REG|code|E12000001")
  )
)
eesyapi:::todf_geographies(
  data.frame(location = c("REG|code|E12000001", "REG|code|E12000001"))
)
eesyapi:::todf_geographies(
  data.frame(
    geographic_level = c("REG", "LA"),
    location = c("REG|code|E12000001", "REG|code|E12000001")
  )
)
eesyapi:::todf_geographies(
  data.frame(geographic_level = c("NAT", "REG"))
)
}
\keyword{internal}
